src/main/java/com/nqbao/project/service/UserService.java
package com.nqbao.project.service;

import com.nqbao.project.model.User;
import com.nqbao.project.model.UserDTO;
import com.nqbao.project.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.cache.annotation.Cacheable;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class UserService extends GenericService<User, Integer, UserDTO> implements UserDetailsService {

    private final UserRepository userRepository;

    @Autowired
    private PasswordEncoder passwordEncoder;

    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        super(userRepository,
                "user",
                UserDTO::new,
                User::new);
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    @Override
    public UserDTO create(UserDTO userDTO) throws NoSuchFieldException, IllegalAccessException {
        userDTO.setPassword(passwordEncoder.encode(userDTO.getPassword()));
        return super.create(userDTO);
    }

    @Cacheable(value = "user", key = "#username")
    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        return userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException(username));
    }

}
